bench(holon-openai): holon-0811-runtime-test-module-split
