+++ src/runtime/tests/mod.rs
@@ -1,0 +1,6 @@
+mod support;
+mod workspace;
+mod work_items;
+mod agent_and_tools;
+mod turns;
+mod runtime_state;
